+++ src/main/java/nextstep/favorite/application/FavoriteService.java
@@ -9,7 +9,7 @@ import nextstep.member.domain.LoginMember;
 import nextstep.member.domain.Member;
 import nextstep.subway.application.PathService;
 import nextstep.subway.application.dto.PathResponse;
-import nextstep.subway.domain.PathType;
+import nextstep.subway.domain.path.PathType;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
