+++ src/main/java/nextstep/subway/domain/path/PathType.java
@@ -1,4 +1,4 @@
-package nextstep.subway.domain;
+package nextstep.subway.domain.path;
 
 public enum PathType {
     DURATION, DISTANCE;
