+++ src/main/java/nextstep/subway/ui/PathController.java
@@ -2,7 +2,7 @@ package nextstep.subway.ui;
 
 import nextstep.subway.application.PathService;
 import nextstep.subway.application.dto.PathResponse;
-import nextstep.subway.domain.PathType;
+import nextstep.subway.domain.path.PathType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
