+++ src/test/java/nextstep/subway/unit/FareTest.java
@@ -1,6 +1,6 @@
 package nextstep.subway.unit;
 
-import nextstep.subway.domain.pathfinder.Fare;
+import nextstep.subway.domain.path.Fare;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
