+++ src/test/java/nextstep/subway/unit/PathServiceMockTest.java
@@ -7,9 +7,9 @@ import nextstep.subway.application.StationService;
 import nextstep.subway.application.dto.PathResponse;
 import nextstep.subway.application.dto.StationResponse;
 import nextstep.subway.domain.Line;
-import nextstep.subway.domain.PathType;
+import nextstep.subway.domain.path.PathType;
 import nextstep.subway.domain.Station;
-import nextstep.subway.domain.pathfinder.ShortestDistancePathFinder;
+import nextstep.subway.domain.path.ShortestDistancePathFinder;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
