+++ src/test/java/nextstep/subway/unit/ShortestDistancePathFinderTest.java
@@ -3,8 +3,8 @@ package nextstep.subway.unit;
 import nextstep.subway.application.dto.PathResponse;
 import nextstep.subway.domain.Line;
 import nextstep.subway.domain.Section;
-import nextstep.subway.domain.pathfinder.PathFinder;
-import nextstep.subway.domain.pathfinder.ShortestDistancePathFinder;
+import nextstep.subway.domain.path.PathFinder;
+import nextstep.subway.domain.path.ShortestDistancePathFinder;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.web.server.ResponseStatusException;
